+++ Machina/Control.cs
@@ -1281,7 +1281,11 @@ namespace Machina
             this.parentRobot.OnActionCompleted(e);
         }
 
-        
+        internal void RaiseMotionCursorUpdated()
+        {
+            MotionCursorUpdatedArgs e = new MotionCursorUpdatedArgs(this.GetCurrentPosition(), this.GetCurrentRotation(), this.GetCurrentAxes(), this.GetCurrentExternalAxes());
+            this.parentRobot.OnMotionCursorUpdated(e);
+        }
 
     }
 }