+++ Machina/Drivers/Communication/TCPCommunicationManagerUR.cs
@@ -397,15 +397,9 @@ namespace Machina.Drivers.Communication
             
             _receivedIDs.Add(id);
             this._motionCursor.ApplyActionsUntilId(id);
-            this._parentDriver.parentControl.parentRobot.OnMotionCursorUpdated(EventArgs.Empty);
 
-            //Action lastAction = this._motionCursor.GetLastAction();
-            //int pendingWrite = this._writeCursor.ActionsPendingCount();
-            //int pendingBuffer = this._motionCursor.ActionsPendingCount();
-            //ActionCompletedArgs e = new ActionCompletedArgs(lastAction, pendingWrite + pendingBuffer, pendingBuffer);
-            //this._parentDriver.parentControl.parentRobot.OnActionCompleted(e);
-
-            // All the above moved to Control so that it isn't necessary to handle the event in each CommunicationMnager...
+            // Raise appropriate events
+            this._parentDriver.parentControl.RaiseMotionCursorUpdated();
             this._parentDriver.parentControl.RaiseActionCompletedEvent();
 
             return true;